+++ src/main/java/com/coniverse/dangjang/domain/user/dto/response/DuplicateNicknameResponse.java
@@ -1,4 +1,4 @@
-package com.coniverse.dangjang.domain.user.dto;
+package com.coniverse.dangjang.domain.user.dto.response;
 
 public record DuplicateNicknameResponse(Boolean duplicate) {
