+++ src/test/java/com/coniverse/dangjang/fixture/SignUpFixture.java
@@ -3,7 +3,7 @@ package com.coniverse.dangjang.fixture;
 import java.time.LocalDate;
 import java.util.List;
 
-import com.coniverse.dangjang.domain.user.dto.SignUpRequest;
+import com.coniverse.dangjang.domain.user.dto.request.SignUpRequest;
 
 /**
  * 테스트에서 필요한 Request를 반환해주는 메소드를 가진 클래스
